requirements.txt
fastf1
pygame
requests
pandas
numpy
imageio
pygithub
python-dotenv
